requirements.txt
# API local
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3

# ETL
pandas==2.2.3
requests==2.32.3

# IA Generativa
groq==0.13.0

# Utilitários
python-dotenv==1.0.1

# Testes
pytest==8.3.4
httpx==0.28.1        # TestClient do FastAPI depende do httpx
